Adhere to clean architecture and SOLID by reorganize view package: create 3 view classes for 3 potential screens. Add a new [interface_adaptor] package, move the temporary Start class to there.
