+++ src/main/java/interface_adaptor/Start.java
@@ -1,7 +1,9 @@
-package main.java.view;
+package main.java.interface_adaptor;
+import main.java.view.StartFrame;
 
 public class Start {
     public static void main(String[] args) {
         new StartFrame();
     }
 }
+
