+++ src/main/java/interface_adaptor/ViewManager.java
@@ -1,0 +1,4 @@
+package main.java.interface_adaptor;
+
+public class ViewManager {
+}
