+++ src/main/java/view/SavingView.java
@@ -1,0 +1,4 @@
+package main.java.view;
+
+public class SavingView {
+}
