+++ src/main/java/view/SearchView.java
@@ -1,0 +1,4 @@
+package main.java.view;
+
+public class SearchView {
+}
