+++ src/main/java/view/SettingsView.java
@@ -1,0 +1,4 @@
+package main.java.view;
+
+public class SettingsView {
+}
